+++ ChangeLog.md
@@ -1,3 +1,8 @@
 # Changelog for project-euler
 
-## Unreleased changes
+## Version 0.1.0.1
+- moved `pascal` (Polynomials -> NumbersExtra) and removed duplicate `choose`
+- updated .gitignore to exclude recent problems
+
+## Version 0.1.0.0
+- Questions 1-61 and 709